refactor(core): extract CommandRegistry into ts-command-registry crate (#354)
Foundational-вынос под Фазу 2 (#354): trait `CommandRegistry` + макрос `register_commands!`
из монолита `src-tauri/src/command_registry.rs` в новый крейт `ts-command-registry`
(только tauri). Разблокирует вынос media/montage_planner/security (их `*CommandRegistry`
импортируют этот trait).

- crates/ts-command-registry: lib.rs (бывш. command_registry.rs) + перенесённый тест-файл.
- Монолит: command_registry.rs -> ре-экспорт-шим `pub use ts_command_registry::CommandRegistry;`
  (5 потребителей резолвятся без правок). + path-dep в src-tauri/Cargo.toml.
- doc-пример макроса помечен ```ignore (иллюстративный).

Проверки: cargo test -p ts-command-registry -> 15 passed; cargo check --lib (src-tauri) ->
0 ошибок.
