+++ crates/ts-command-registry/src/command_registry_tests.rs
